MULE-4155 missing call to client.completePendingCommand in postProcess
git-svn-id: https://svn.codehaus.org/mule/branches/mule-2.x@14009 bf997673-6b11-0410-b953-e057580c5b09

+++ transports/ftp/src/main/java/org/mule/transport/ftp/FtpMessageReceiver.java
@@ -201,6 +201,11 @@ public class FtpMessageReceiver extends AbstractPollingMessageReceiver
 
     protected void postProcess(FTPClient client, FTPFile file, MuleMessage message) throws Exception
     {
+        if (!client.completePendingCommand())
+        {
+            throw new IOException(MessageFormat.format("Failed to complete a pending command. Retrieveing file {0}. Ftp error: {1}",
+                                                       file.getName(), client.getReplyCode()));
+        }
         if (!client.deleteFile(file.getName()))
         {
             throw new IOException(MessageFormat.format("Failed to delete file {0}. Ftp error: {1}",
